myproject/requirements.txt
alabaster==0.7.12
apparmor==3.0.4
appdirs==1.4.4
apturl==0.5.2
asarPy==1.0.1
asgiref==3.8.1
astroid==2.9.3
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.6.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
black==21.10b0
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cloudpickle==2.0.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diff-match-patch==20200713
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
docutils==0.17.1
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gyp==0.1
helpdev==0.7.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
imagesize==1.3.0
importlib-metadata==4.6.4
intervaltree==3.0.2
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
isort==5.6.4
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
LibAppArmor==3.0.4
libvirt-python==8.0.0
lockfile==0.12.2
logilab-common==1.8.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mercurial==6.1.1
mock==4.0.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
mypy-extensions==0.4.3
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
numpy==1.21.5
numpydoc==1.2
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pathspec==0.9.0
pbr==5.8.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==2.12.2
pyls-black==0.4.6
pyls-spyder==0.4.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
PyQtWebEngine==5.15.5
pyRFC3339==1.1
pyrsistent==0.18.1
pysvn==1.9.15
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
QDarkStyle==2.8.1
QtAwesome==1.1.1
qtconsole==5.2.2
QtPy==2.0.0
reportlab==3.6.8
requests==2.25.1
roman==3.3
rope==0.22.0
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.1.0
soupsieve==2.3.1
Sphinx==4.3.2
spyder==4.2.1
spyder-kernels==1.10.2
sqlparse==0.5.0
sympy==1.9
systemd-python==234
testpath==0.5.0
textdistance==4.2.2
three-merge==0.1.1
toml==0.10.2
tomli==1.2.2
tornado==6.1
traitlets==5.1.1
ttystatus==0.38
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.1.0
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
wurlitzer==3.0.2
xdg==5
xkit==0.0.0
yapf==0.32.0
zipp==1.0.0
